Updated structures to 10.1.5 * ref https://github.com/TrinityCore/TrinityCore/commit/0cea730fa23473a85c47451c3bd13df816f2b6e4

+++ WowPacketParser/Store/Objects/QuestTemplate.cs
@@ -250,7 +250,7 @@ namespace WowPacketParser.Store.Objects
         public uint? AreaGroupID;
 
         [DBFieldName("TimeAllowed")]
-        public uint? TimeAllowed;
+        public ulong? TimeAllowed;
 
         [DBFieldName("AllowableRaces", TargetedDatabaseFlag.TillCataclysm)]
         public RaceMask? AllowableRaces;
